Waterfall starts to work as expected, coolio!
