+++ functions.js
@@ -24,50 +24,34 @@ exports.myIncomingCommand = function(rows) // Check if command exists.
 
 exports.myIncomingProduct = function(rows) // Check, if product exists.
     {
-        if (rows.lenght == 0) {
-            return 'Product ' + data.product + ' not available! Please try /TRD PRODUCTS to see, what is available.';
+        if (rows.length == 0) {
+            return 'Product ' + _.values(data.data.product) + ' not available! Please try /TRD PRODUCTS to see, what is available.';
         } else return null;
     }
 
-
-
-
-exports.displayProducts = function(rows, fields, context) // Special command to display available products on market
-    {
+exports.showBidPrices = function(rows) {
         var result = '';
         _.forEach(rows, function(value) {
-            result += value.product_name + ', ';
+            result += value.price + ', ';
         });
-        context.succeed('Available products: ' + result.toUpperCase());
-        return;
-    }
+        return 'You can SELL ' + _.values(data.data.product) + ' for these prices: ' + result;
+}
 
-exports.incomingOrder = function(err, data, context) {
+exports.showAskPrices = function(rows) {
+        var result = '';
+        _.forEach(rows, function(value) {
+            result += value.price + ', ';
+        });
+        return 'You can BUY ' + _.values(data.data.product) + ' for these prices: ' + result;
+}
+
+exports.myOrder = function(err, data, context) {
     if (err !== null)
         context.fail(err);
     else {
         console.log('Order successfully inserted! ' + data.command + ' ' + data.product + ' ' + data.price);
     }
 }
-
-exports.incomingCommand = function(commandRows, fields, context) // Check if command exists.
-    {
-        if (commandRows == null) {
-            context.succeed('Order type ' + _.values(data.data) + ' not available! Please try /TRD HELP first!');
-        }
-        console.log(commandRows);
-    }
-
-exports.incomingProduct = function(err, productRows, context) // Check, if product exists.
-    {
-        if (err !== null)
-            context.fail(err);
-        else {
-            if (productRows == null) {
-                if (context) context.succeed('Product ' + data.product + ' not available! Please try /TRD PRODUCTS to see, what is available.');
-            }
-        }
-    }
 
 exports.countingOrders = function(err, countRows, context) {
     exports.totalOrders = totalOrders;
@@ -96,14 +80,3 @@ exports.deleteHigh = function(err, delRows, context) {
     }
 }
 
-exports.showPrices = function(err, dataRows, context) {
-    if (err !== null)
-        context.fail(err);
-    else {
-        var result = '';
-        _.forEach(dataRows, function(value) {
-            result += value.price + ', ';
-        });
-        context.succeed('You can ' + context.data.command + ' ' + context.data.product + ' for these prices: ' + result);
-    }
-}
