+++ test_events/event_unique.json
@@ -1,3 +1,3 @@
 {
-  "text"   : "buy bram"
+  "text"   : "buy stone"
 }
